requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
torch>=2.2
torchvision>=0.17
scikit-learn>=1.3
Pillow>=10.0
tqdm>=4.66